Version v.1.0.1

+++ app/api/auth/adminuser/[id]/route.ts
@@ -1,7 +1,32 @@
 import { NextRequest, NextResponse } from "next/server";
 import pool from "@/lib/db";
 import { hashPassword } from "@/lib/bc";
-import { ResultSetHeader } from "mysql2/promise";
+import { ResultSetHeader, RowDataPacket } from "mysql2/promise";
+
+// Handler for GET request to fetch a user by ID
+export async function GET(
+    req: NextRequest,
+    { params }: { params: { id: string } }
+) {
+    try {
+        const { id } = params;
+        
+        const [userRows] = await pool.query<RowDataPacket[]>(
+            'SELECT * FROM admin_users WHERE id = ?',
+            [id]
+        );
+
+        if (userRows.length === 0) {
+            return NextResponse.json({ message: 'Usuario no encontrado' }, { status: 404 });
+        }
+
+        const user = userRows[0];
+        return NextResponse.json(user, { status: 200 });
+    } catch (error) {
+        const errorMessage = (error as Error).message;
+        return NextResponse.json({ message: errorMessage }, { status: 500 });
+    }
+}
 
 // Handler for POST request to create a new admin user
 export async function POST(req: NextRequest) {
@@ -67,12 +92,24 @@ export async function DELETE(
     try {
         const { id } = params;
 
+        // Obtener la información del usuario antes de eliminarlo
+        const [userRows] = await pool.query<RowDataPacket[]>(
+            'SELECT * FROM admin_users WHERE id = ?',
+            [id]
+        );
+
+        if (userRows.length === 0) {
+            return NextResponse.json({ message: 'Usuario no encontrado' }, { status: 404 });
+        }
+
+        const user = userRows[0];
+
         await pool.query<ResultSetHeader>(
             "DELETE FROM admin_users WHERE id = ?",
             [id]
         );
 
-        return NextResponse.json({}, { status: 204 });
+        return NextResponse.json({ message: `El usuario ${user.name} ${user.lastname} ha sido eliminado.` }, { status: 200 });
     } catch (error) {
         const errorMessage = (error as Error).message;
         return NextResponse.json({ message: errorMessage }, { status: 500 });
